[#data] Added Test Cases and handled last segement addition
